requirements.txt
# Local LLM Project Manager -- Python dependencies
#
# Python >= 3.9 required (3.13 recommended).
#
# Core pipeline (main.py + mcp/) needs NOTHING beyond the standard library.
#
# Web UI (app.py)
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
#
# Jira / Confluence connectors (connectors/)
requests>=2.28.0
python-dotenv>=1.0.0
